requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
redis==5.0.1
APScheduler==3.10.4
